+++ frontend/views/profiles/show.jsx
@@ -1,7 +1,7 @@
 // DEPENDENCIES
 const React = require('react');
 const Def = require('../default');
-const DefPro = require('./defaultPro')
+// const DefPro = require('./defaultPro')
 
 // SHOWS AN INDIVIDUAL PROFILE
 function show (data) {
@@ -21,7 +21,7 @@ function show (data) {
                         <button className="btn btn-info characterButton">{c.name}</button>
                     </div>
                     <div>
-                        <img class="characterAvatar" src={c.avatar} alt={c.name}></img>
+                        <img className="characterAvatar" src={c.avatar} alt={c.name}></img>
                     </div>
                     <div>
                         <h5>Profile</h5>
@@ -44,9 +44,16 @@ function show (data) {
                 <p>{data.profile.info}</p>
 
                 {/* BUTTONS */}
+                {/* EDIT */}
                 <a href={`/profiles/${data.profile.id}/edit`} className="btn btn-warning">
                     Edit
                 </a>
+                {/* NEW CHARACTER  */}
+                {/* hiding this button for now because it's having issues */}
+                {/* <a href={`/profiles/${data.profile.id}/characters/new`} className="btn btn-success">
+                    Submit Character
+                </a> */}
+                {/* DELETE */}
                 <form method="POST" action={`/profiles/${data.profile.id}?_method=DELETE`}>
                     <button type="submit" className="btn btn-danger">
                         Delete
@@ -56,6 +63,42 @@ function show (data) {
                 {/* CHARACTERS */}
                 <h2>Characters</h2>
                 <div className="formattedCharacters">{characters}</div>
+
+
+                <h1>Submit New Character</h1>
+                <form method="POST" action={`/profiles/${data.profile.id}/character`}>
+                    <container className='registration'>
+                        {/* CHARACTER NAME */}
+                        <div className='form-group'>
+                            <label htmlFor='name'>Character Name </label>
+                            <input className='form-control' type='text' placeholder='Character Name' id='name' name='name' required aria-describedby="required-description" />
+                        </div>
+                        {/* AVATAR */}
+                        <div className='form-group'>
+                            <label htmlFor='avatar'>Avatar</label>
+                            <input className='form-control' type='text' placeholder='Avatar Link' id='avatar' name='avatar'/>
+                        </div>
+                        {/* PROFILE */}
+                        <div className='form-group'>
+                            <label htmlFor='profile'>Profile</label>
+                            {/* <textarea cols='3' rows='3'
+                            className='form-control' id='profile' name='profile'
+                            /> */}
+                            <input className='form-control' type='text' placeholder='profile' id='profile' name='profile'/>
+                        </div>
+                        {/* PERMISSIONS */}
+                        <div className='form-group'>
+                            <label htmlFor='permissions'>Permissions</label>
+                            {/* <textarea cols='3' rows='3'
+                            className='form-control' id='permissions' name='permissions'
+                            /> */}
+                            <input className='form-control' type='text' placeholder='permissions' id='permissions' name='permissions'/>
+                        </div>
+                        {/* SUBMIT BUTTON */}
+                        <input type='submit' className='submit-button' value='Submit Character'/>                        
+                    </container>
+                    
+                </form>
             </main>
         </Def>
     )